app/requirements.txt
adal==1.2.4
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
idna==2.9
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
